requirements.txt
arrow==1.2.3
asgiref==3.6.0
backports.zoneinfo==0.2.1
blessed==1.19.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-picklefield==3.1
django-q==1.3.9
idna==3.4
PyJWT==2.6.0
python-dateutil==2.8.2
redis==3.5.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
wcwidth==0.2.6
